+++ src/utils/BaseEvent.ts
@@ -1,5 +1,4 @@
-
-import DiscordClient from '../../client/client';
+import DiscordClient from './client/client';
 
 export default abstract class BaseEvent {
   constructor(private name: string) { }
